requirements.txt
discord.py 
python-dotenv
mysql-connector-python